refactor: extract shared PostForm component from Create and Edit pages
The form fields, toggleCategory logic, and validation error display
were nearly identical. Now uses a shared PostForm component; page-level
concerns (breadcrumbs, autosave) remain in the page components.
